Pages for Second Sprint Created
The page that has the manager options that they can use has been created with functionality. The page for recording a new vehicles information has been created and interacts with the database. It will upload a new vehicle to the database, however it doesn't input a vehicles sale price or availability. This can be added if needed but as of now its not implemented.

Header comments were also added to every file that is currently present in the project.

The FXML files for every panel were created and added to this commit except one that is for displaying the information on a vehicle that is searched. The controller classes for these FXML files haven't been created however, these will come later.