Refactored Code
This commit contains code that has been refactored and put in individual folders so that everything is more efficient to find and get access to.

The FXML files are in the resources folder where they should be.

When creating a new controller class for a panel you need to specify in the JavaFXController portion org.dealership.controllerClasses.*insert controller class name* .

Header comments were also put on every file.