Controller classes, Project reformat, Commission page
I went through and organized all of the files to be in a much more efficient way of being stored and just to be found.

Specific file types will be in their own respective folders inside the project.

I also created the missing controller classes for the fxml files that the employee has access to and added in the fxml code to said controller classes.

You can access the pages and go back and forth between them as you wish with no issue, as of now.

When using it you may need to run some of the new SQL statements in the testingSQLquery file that is in the project, you'll have to make sure you know what is in your sql database already and all.

There was also some refactoring done in the code to make it just easier to read.