Trade in page, bugs fixed
This commit contains code that gave the trade in vehicle page functionality for inputting a new trade in vehicle into the database.

I also fixed the bug where you could login to either an employee or manager page regardless of the password used. It was only checking user ID and not password like intended.

I also fixed the bug where on the order vehicle page, depending on if you were logged in as a regular employee or manager it would always go back to employee options page.