New Unit Tests, New validation Methods
This commit contains the tests for the Commission amount for an employee and the corresponding input validations.

The same goes for the recording customer information as well.

Same file path can be used that was listed in the last commit.

Some files have been deleted as they weren't needed and were taking up space, this hasn't affected anything in terms of the system working.

Comments and all have been added to some things as well.